feat: remove need for explicit 'local' declaration
That is, when a variable is referenced, space is created for it
automatically. This simplifies having to deal with how to handle
'local' as a proper unary operator that should theoretically be able
to handle _any_ input expression. We also no longer need to deal with
undeclared local variables.

A good sign is that the original register tests now fall through with
only minor changes (for example, double-dereferencing still works.)

Maybe next I'll work on booleans and such. Conditionals and loops -
and especially function calls as a kind of final boss - will be a bit
harder, since they'll ultimately involving labels and goto at least
internally. We'll see.