feat: allow for nested quotation
This allows for what are poor-man's nested functions, e.g.

x <- { y <- { 10 * 10 } ; 12 }
=> <heap addr>

call x
=> 12

call y
=> 100

An awkward feature arises where, since all local variables are
initialized to 0, and 0 in this instance is a valid call-address,
'call y' works in invoking the code that here explicitly gets assigned
to x.